// src/enums/RioInputEnum.ts
export enum RioInputEnum {
  PRIVACY = "Privacy",
  MARKET_ADVANTAGE = "Market Advantage",
  LONG_TERM_RESILIENCE = "Long Term Resilience",
  LONG_TERM_SUSTAINABILITY = "Long Term Sustainability",
  LEGAL_REQUIREMENTS = "Legal Requirements",
  INNOVATION = "Innovation",
  OTHER_RISKS = "Other Risks",
}